+++ Controllers/InvoiceController.cs
@@ -10,7 +10,6 @@ namespace FullStackApi.Controllers;
 [Route("api/[controller]")]
 public class InvoiceController : ControllerBase
 {
-
   private readonly FullStackApiContext _context;
 
   public InvoiceController(FullStackApiContext context)
@@ -37,11 +36,11 @@ public class InvoiceController : ControllerBase
   [HttpPost]
   public async Task<IActionResult> Post(Invoice newInvoice)
   {
+    if (newInvoice == null) return BadRequest();
     _context.Invoices.Add(newInvoice);
     await _context.SaveChangesAsync();
 
     Invoice invoice = await _context.Invoices.FindAsync(newInvoice.Id);
-
     return Ok(invoice);
   }
 
@@ -49,6 +48,7 @@ public class InvoiceController : ControllerBase
   public async Task<IActionResult> Delete(int id)
   {
     Invoice invoice = await _context.Invoices.FindAsync(id);
+    if (invoice == null) return BadRequest();
     _context.Invoices.Remove(invoice);
     await _context.SaveChangesAsync();
     return NoContent();
@@ -57,6 +57,7 @@ public class InvoiceController : ControllerBase
   [HttpPut]
   public async Task<IActionResult> Put(Invoice editedInvoice)
   {
+    if (editedInvoice == null) return BadRequest();
     _context.Entry(editedInvoice).State = EntityState.Modified;
     await _context.SaveChangesAsync();
     return NoContent();
